requirements.txt
attrs==23.2.0
blessed==1.20.0
certifi==2024.2.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
editor==1.6.6
flatbuffers==24.3.25
humanfriendly==10.0
idna==3.7
imageio==2.34.1
inquirer==3.2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lazy_loader==0.4
llvmlite==0.42.0
mpmath==1.3.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
onnxruntime==1.18.0
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.1
protobuf==5.26.1
PyMatting==1.1.12
readchar==4.1.0
referencing==0.35.1
rembg==2.0.56
requests==2.31.0
rpds-py==0.18.1
runs==1.2.2
scikit-image==0.23.2
scipy==1.13.0
six==1.16.0
sympy==1.12
tifffile==2024.5.10
tqdm==4.66.4
urllib3==2.2.1
wcwidth==0.2.13
xmod==1.8.1
